demo phan warehouse cho cac parcel reqquest

+++ share-a-ride-taxi/openerp-resource-server/src/main/java/openerp/openerpresourceserver/entity/Warehouse.java
@@ -36,10 +36,10 @@ public class Warehouse {
     private String addressNote;
 
     @Column(name = "latitude")
-    private BigDecimal lat;
+    private Double lat;
 
     @Column(name = "longitude")
-    private BigDecimal lon;
+    private Double lon;
 
     @Column(name = "created_at")
     private LocalDateTime createdAt;
